# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Artifact Hub
  description: "Find, install and publish Kubernetes packages"
  version: 0.20.0
  contact:
    name: Artifact Hub support
    url: https://cloud-native.slack.com/channels/artifact-hub
servers:
  - url: "/api/v1"
...
paths:
...
  "/orgs/{orgName}/user-allowed-actions":
    get:
      tags:
        - Organizations
      security:
        - ApiKeyId: []
          ApiKeySecret: []
      summary: >-
        Get actions which user is allowed to perform in the provided
        organization
      description: >-
        Get actions which user is allowed to perform in the provided
        organization
      operationId: getAllowedActions
      parameters:
        - $ref: "#/components/parameters/OrgNameParam"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuthorizerAction"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    ApiKeyId:
      type: apiKey
      in: header
      name: X-API-KEY-ID
    ApiKeySecret:
      type: apiKey
      in: header
      name: X-API-KEY-SECRET
  schemas:
    AuthorizerAction:
      type: string
      enum:
        - all
        - addOrganizationMember
        - addOrganizationRepository
        - deleteOrganizationMember
        - deleteOrganizationRepository
        - getAuthorizationPolicy
        - transferOrganizationRepository
        - updateAuthorizationPolicy
        - updateOrganization
        - updateOrganizationRepository
      description: >
        Authorization policy action:

        * `all` - All actions

        * `addOrganizationMember` - Add member to organization

        * `addOrganizationRepository` - Add repository to organization

        * `deleteOrganizationMember` - Delete member from organization

        * `deleteOrganizationRepository` - Delete repository from organization

        * `getAuthorizationPolicy` - Get authorization policy

        * `transferOrganizationRepository` - Transfer repository from
        organization

        * `updateAuthorizationPolicy` - Update authorization policy

        * `updateOrganization` - Update organization

        * `updateOrganizationRepository` - Update repository from organization
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: error details
...
  parameters:
...
    OrgNameParam:
      in: path
      name: orgName
      schema:
        type: string
        example: org1
      required: true
      description: Organization name
...
  responses:
...
    Forbidden:
      description: The user does not have permission to perform the requested operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: >-
        The server encountered an unexpected condition that prevented it from
        fulfilling the request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    TooManyRequests:
      description: The user has sent too many requests in a given amount of time
    UnauthorizedError:
      description: Valid authentication credentials not provided
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...